refactor accessibility mode logic into proper files
